[MOD] Moved copy gameobject code
